Fixed broken "non blocking" mode backport from master

+++ src/main/java/com/github/shyiko/mysql/binlog/BinaryLogClient.java
@@ -621,7 +621,7 @@ public class BinaryLogClient implements BinaryLogClientMXBean {
                     throw new ServerException(errorPacket.getErrorMessage(), errorPacket.getErrorCode(),
                         errorPacket.getSqlState());
                 }
-                if (marker == (byte) 0xFE && !blocking) {
+                if (marker == 0xFE && !blocking) {
                     completeShutdown = true;
                     break;
                 }
